Add Elephant Factor metric
The Elephant Factor metric implementation is based on Pandas.
The architecture is divided in:
 - broomstick.data: access data layer, focused on ES by the time
   being.
 - broomstick.metrics: service layer providing metrics.
 - broomstick.core: common stuff needed by the data and service
   layers.

The notebooks folder contains a Jupyter notebook that uses
broomstick. Notebooks are the preferred way to use broomstick
library (up to now). We still need to define the visualization
part and decide whether broomstick is able to plot charts or
not. In this version it is, however there is a flag to not
plot anything if needed.

Tests have been implemented for all the available methods.

Signed-off-by: Alberto Pérez García-Plaza <alpgarcia@bitergia.com>